bk-plugin-framework/dev-requirements.txt

werkzeug>=2.0.0,<3.0.0
pydantic>1.0,<2.0
jsonschema>=2.5.1,<5.0

# services
Django>=2.2,<3.0
djangorestframework==3.12.4
drf-yasg==1.20.0
apigw-manager==0.1.8

# runtime
celery==3.1.25

# test
pytest==6.2.4
tox==3.24.0
tox-pyenv==1.1.0
coverage==5.5
pytest-django==4.4.0

# code check
flake8
black